+++ packages/liveblocks-schema/src/checker/index.ts
@@ -1,4 +1,4 @@
-import original_didyoumean from "didyoumean";
+import { didyoumean as dym } from "../lib/didyoumean";
 
 import type {
   Definition,
@@ -24,13 +24,50 @@ const TYPENAME_REGEX = /^[A-Z_]/;
 
 // TODO Ideally _derive_ this list of builtins directly from the grammar
 // instead somehow?
-const BUILTINS = ["String", "Int", "Float", "Boolean"];
+const BUILTINS = ["String", "Int", "Float", "Boolean"] as const;
+
+function suggest_general(value: string, alternatives: string[]): string[] {
+  // Never suggest "Storage"
+  alternatives = alternatives.filter((key) => key !== "Storage");
+
+  const suggestions = dym(value, alternatives);
+
+  // Special hack:
+  // It can be expected that people will try to put "number" in as a type,
+  // because that's TypeScript's syntax. If there is no custom type name found
+  // that closely matches this typo, then try to suggest one more thing to
+  // nudge them. (But only if Float and Int are already legal suggestions.)
+  if (
+    suggestions.length === 0 &&
+    alternatives.includes("Float") &&
+    alternatives.includes("Int")
+  ) {
+    if (/^num(ber)?$/i.test(value)) {
+      return ["Float", "Int"];
+    }
+  }
+
+  return suggestions;
+}
+
+function didyoumeanify(message: string, alternatives: string[]): string {
+  if (alternatives.length === 0) {
+    return message;
+  }
+  if (message.endsWith(".")) {
+    message = message.slice(0, -1);
+  }
+  return `${message}. Did you mean ${alternatives.map(quote).join(" or ")}?`;
+}
 
 /**
  * Reserve these names for future use.
  */
 const RESERVED_TYPENAMES_REGEX = /^Live|^(Presence|Array)$/i;
 
+/**
+ * Helper constructs for use during "checking" phase.
+ */
 class Context {
   errorReporter: ErrorReporter;
 
@@ -46,6 +83,22 @@ class Context {
   // field definitions. As such, these types can only be used by other Live
   // types. Defined during the first checkDocument pass.
   liveOnlyTypes: Set<string>;
+
+  readonly suggestors = {
+    objectTypeName: (name: string): string[] =>
+      suggest_general(
+        name,
+        Array.from(this.registeredTypes)
+          .filter(([, def]) => def._kind === "ObjectTypeDefinition")
+          .map(([key]) => key)
+      ),
+
+    typeNameOrBuiltIn: (name: string): string[] =>
+      suggest_general(name, [
+        ...Array.from(this.registeredTypes.keys()),
+        ...BUILTINS,
+      ]),
+  };
 
   constructor(errorReporter: ErrorReporter) {
     this.errorReporter = errorReporter;
@@ -147,89 +200,61 @@ function checkTypeName(node: TypeName, context: Context): void {
   }
 }
 
-/**
- * Wrap didyoumean to avoid dealing with a million different possible output
- * values.
- */
-function didyoumean(value: string, alternatives: string[]): string[] {
-  const output = original_didyoumean(value, alternatives);
-  if (!output) {
-    return [];
-  } else if (Array.isArray(output)) {
-    return output;
+function checkTypeNameExists(
+  node: TypeName,
+  context: Context,
+  suggestor: (name: string) => string[]
+): void {
+  if (context.registeredTypes.has(node.name)) {
+    return;
+  }
+  context.report(
+    didyoumeanify(`Unknown type ${quote(node.name)}`, suggestor(node.name)),
+    node.range
+  );
+}
+
+function checkTypeRefTarget(node: TypeRef, context: Context): void {
+  if (node.asLiveObject) {
+    checkTypeNameIsObjectType(node.ref, context);
+    checkTypeNameExists(node.ref, context, context.suggestors.objectTypeName);
   } else {
-    return [output];
+    checkTypeNameExists(
+      node.ref,
+      context,
+      context.suggestors.typeNameOrBuiltIn
+    );
   }
 }
 
-function checkTypeRefTargetExists(node: TypeRef, context: Context): void {
-  const name = node.ref.name;
-  const typeDef = context.registeredTypes.get(name);
-  if (typeDef !== undefined) {
+function checkTypeNameIsObjectType(node: TypeName, context: Context): void {
+  const def = context.registeredTypes.get(node.name);
+  if (!def) {
+    context.report(
+      didyoumeanify(
+        `Unknown object type ${quote(node.name)}`,
+        context.suggestors.objectTypeName(node.name)
+      ),
+      node.range
+    );
     return;
   }
 
-  // If we land here, it means there's an unknown type reference. Possibly
-  // caused by misspellings or people trying to learn/play with the language.
-  // Let's be friendly to them and assist them with fixing the problem,
-  // especially around common mistakes.
-  let alternatives: string[] = didyoumean(
-    node.ref.name,
-    BUILTINS.concat(Array.from(context.registeredTypes.keys()))
-  );
-
-  if (alternatives.length === 0) {
-    // It can be expected that people will try to put "number" in as a type,
-    // because that's TypeScript's syntax. If there is no custom type name
-    // found that closely matches this typo, then try to suggest one more thing
-    // to nudge them.
-    alternatives = /^num(ber)?$/i.test(name)
-      ? ["Float", "Int"]
-      : [
-          /* no alternatives */
-        ];
-  }
-
-  const suggestion =
-    alternatives.length > 0
-      ? `. Did you mean ${alternatives.map((alt) => quote(alt)).join(" or ")}?`
-      : "";
-
-  context.report(`Unknown type ${quote(name)}` + suggestion, node.ref.range);
-}
-
-function checkLiveObjectPayloadIsObjectType(
-  typeRef: TypeRef,
-  context: Context
-): void {
   // Check that the payload of a LiveObject type is an object type
-  if (
-    typeRef.asLiveObject &&
-    context.registeredTypes.get(typeRef.ref.name)?._kind !==
-      "ObjectTypeDefinition"
-  ) {
-    const alternatives: string[] = didyoumean(
-      typeRef.ref.name,
-      Array.from(context.registeredTypes)
-        .filter(([, def]) => def._kind === "ObjectTypeDefinition")
-        .map(([key]) => key)
-    );
-
+  if (def._kind !== "ObjectTypeDefinition") {
     context.report(
-      `Type ${quote(typeRef.ref.name)} is not an object type` +
-        (alternatives.length > 0
-          ? `. Did you mean ${alternatives.map(quote).join(" or ")}?`
-          : ""),
-      typeRef.ref.range
+      didyoumeanify(
+        `Type ${quote(node.name)} is not an object type`,
+        context.suggestors.objectTypeName(node.name)
+      ),
+      node.range
     );
     return undefined;
   }
 }
 
 function checkTypeRef(ref: TypeRef, context: Context): void {
-  checkLiveObjectPayloadIsObjectType(ref, context);
-
-  checkTypeRefTargetExists(ref, context);
+  checkTypeRefTarget(ref, context);
 
   //
   // For each definition, first ensure that it and annotate whether or not they